Split water_conservation.rs into water_conservation/ subdirectory
Split crates/simulation/src/water_conservation.rs (640 lines) into focused modules
under water_conservation/ for better maintainability:

- constants.rs (36 lines): All policy reduction/cost constants
- types.rs (49 lines): WaterConservationState resource + Default impl
- calculations.rs (66 lines): Pure helper functions for demand/cost/savings
- systems.rs (64 lines): Bevy system + plugin registration
- tests.rs (417 lines): All unit tests
- mod.rs (9 lines): Thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
