+++ crates/simulation/src/water_conservation/tests.rs
@@ -1,218 +1,8 @@
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::buildings::Building;
-use crate::weather::Weather;
-use crate::SlowTickTimer;
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// Demand reduction from low-flow fixtures (applied to residential buildings only).
-pub const LOW_FLOW_DEMAND_REDUCTION: f32 = 0.20;
-
-/// Demand reduction from xeriscaping (reduced irrigation area, applies to total demand).
-pub const XERISCAPING_DEMAND_REDUCTION: f32 = 0.10;
-
-/// Demand reduction from tiered water pricing (behavioural change, applies to total demand).
-pub const TIERED_PRICING_DEMAND_REDUCTION: f32 = 0.15;
-
-/// Demand reduction from greywater recycling (applies to total demand).
-pub const GREYWATER_DEMAND_REDUCTION: f32 = 0.15;
-
-/// Sewage volume reduction from greywater recycling (greywater reused instead of discharged).
-pub const GREYWATER_SEWAGE_REDUCTION: f32 = 0.30;
-
-/// Demand reduction from rainwater harvesting (effective only when precipitation > 0).
-pub const RAINWATER_DEMAND_REDUCTION: f32 = 0.10;
-
-/// Hard cap on total demand reduction from all combined conservation policies.
-pub const MAX_TOTAL_DEMAND_REDUCTION: f32 = 0.60;
-
-/// Per-building retrofit cost for low-flow fixtures (dollars).
-pub const LOW_FLOW_COST_PER_BUILDING: f64 = 500.0;
-
-/// Per-building retrofit cost for greywater recycling (dollars).
-pub const GREYWATER_COST_PER_BUILDING: f64 = 3000.0;
-
-/// Per-building retrofit cost for rainwater harvesting (dollars).
-pub const RAINWATER_COST_PER_BUILDING: f64 = 1000.0;
-
-/// Base daily water demand per building used for annual savings estimates (gallons).
-/// This is a rough average across building types for estimating conservation savings.
-const BASE_DAILY_DEMAND_PER_BUILDING: f64 = 1200.0;
-
-/// Days in a year for annual savings calculation.
-const DAYS_PER_YEAR: f64 = 365.0;
-
-// =============================================================================
-// Resource
-// =============================================================================
-
-/// City-wide water conservation policy state.
-///
-/// Each boolean field represents an individually toggleable conservation policy.
-/// Other simulation systems (e.g. `water_demand`) read `demand_reduction_pct` to
-/// apply the aggregate reduction. This resource does NOT modify the `Policy` enum;
-/// it is a standalone system with its own policy toggles.
-#[derive(Resource, Debug, Clone, Serialize, Deserialize)]
-pub struct WaterConservationState {
-    /// Low-flow fixtures installed in residential buildings (-20% residential demand).
-    pub low_flow_fixtures: bool,
-    /// Xeriscaping programme active (-10% total demand from reduced irrigation).
-    pub xeriscaping: bool,
-    /// Tiered water pricing in effect (-15% total demand from behavioural change).
-    pub tiered_pricing: bool,
-    /// Greywater recycling enabled (-15% demand AND -30% sewage volume).
-    pub greywater_recycling: bool,
-    /// Rainwater harvesting systems installed (-10% demand when raining).
-    pub rainwater_harvesting: bool,
-    /// Aggregate demand reduction percentage from all active policies (0.0 .. 0.60).
-    pub demand_reduction_pct: f32,
-    /// Sewage volume reduction percentage from greywater recycling (0.0 or 0.30).
-    pub sewage_reduction_pct: f32,
-    /// Cumulative dollar cost of all building retrofits.
-    pub total_retrofit_cost: f64,
-    /// Estimated annual water savings in gallons based on current policies and building count.
-    pub annual_savings_gallons: f64,
-    /// Number of buildings that have been (or would be) retrofitted.
-    pub buildings_retrofitted: u32,
-}
-
-impl Default for WaterConservationState {
-    fn default() -> Self {
-        Self {
-            low_flow_fixtures: false,
-            xeriscaping: false,
-            tiered_pricing: false,
-            greywater_recycling: false,
-            rainwater_harvesting: false,
-            demand_reduction_pct: 0.0,
-            sewage_reduction_pct: 0.0,
-            total_retrofit_cost: 0.0,
-            annual_savings_gallons: 0.0,
-            buildings_retrofitted: 0,
-        }
-    }
-}
-
-// =============================================================================
-// Helper functions (pure, testable)
-// =============================================================================
-
-/// Compute the effective rainwater harvesting reduction given the current
-/// precipitation intensity. Full effectiveness at >= 1.0 in/hr, linearly
-/// scaled below that.
-fn rainwater_effectiveness(precipitation_intensity: f32) -> f32 {
-    if precipitation_intensity <= 0.0 {
-        0.0
-    } else {
-        // Linear scale: full effect at 1.0 in/hr, proportional below.
-        (precipitation_intensity / 1.0).min(1.0) * RAINWATER_DEMAND_REDUCTION
-    }
-}
-
-/// Calculate total demand reduction percentage from the set of active policies
-/// and current precipitation. Result is capped at `MAX_TOTAL_DEMAND_REDUCTION`.
-fn calculate_demand_reduction(state: &WaterConservationState, precipitation: f32) -> f32 {
-    let mut reduction = 0.0_f32;
-
-    if state.low_flow_fixtures {
-        reduction += LOW_FLOW_DEMAND_REDUCTION;
-    }
-    if state.xeriscaping {
-        reduction += XERISCAPING_DEMAND_REDUCTION;
-    }
-    if state.tiered_pricing {
-        reduction += TIERED_PRICING_DEMAND_REDUCTION;
-    }
-    if state.greywater_recycling {
-        reduction += GREYWATER_DEMAND_REDUCTION;
-    }
-    if state.rainwater_harvesting {
-        reduction += rainwater_effectiveness(precipitation);
-    }
-
-    reduction.min(MAX_TOTAL_DEMAND_REDUCTION)
-}
-
-/// Calculate total retrofit cost for all policies that require per-building investment.
-fn calculate_retrofit_cost(state: &WaterConservationState, building_count: u32) -> f64 {
-    let count = building_count as f64;
-    let mut cost = 0.0_f64;
-
-    if state.low_flow_fixtures {
-        cost += LOW_FLOW_COST_PER_BUILDING * count;
-    }
-    if state.greywater_recycling {
-        cost += GREYWATER_COST_PER_BUILDING * count;
-    }
-    if state.rainwater_harvesting {
-        cost += RAINWATER_COST_PER_BUILDING * count;
-    }
-
-    cost
-}
-
-/// Estimate annual water savings in gallons based on demand reduction and building count.
-fn calculate_annual_savings(demand_reduction_pct: f32, building_count: u32) -> f64 {
-    let daily_savings =
-        BASE_DAILY_DEMAND_PER_BUILDING * building_count as f64 * demand_reduction_pct as f64;
-    daily_savings * DAYS_PER_YEAR
-}
-
-// =============================================================================
-// System
-// =============================================================================
-
-/// System: Recalculate water conservation metrics every slow tick.
-///
-/// 1. Counts buildings to determine retrofit scope.
-/// 2. Computes aggregate `demand_reduction_pct` (capped at 0.60).
-/// 3. Computes `sewage_reduction_pct` from greywater policy.
-/// 4. Computes `total_retrofit_cost` from per-building policy costs.
-/// 5. Adjusts rainwater harvesting effectiveness by current precipitation.
-/// 6. Updates estimated `annual_savings_gallons`.
-pub fn update_water_conservation(
-    timer: Res<SlowTickTimer>,
-    weather: Res<Weather>,
-    mut conservation: ResMut<WaterConservationState>,
-    buildings: Query<&Building>,
-) {
-    if !timer.should_run() {
-        return;
-    }
-
-    let building_count = buildings.iter().count() as u32;
-
-    // 1. Demand reduction (precipitation-aware for rainwater harvesting)
-    let precipitation = weather.precipitation_intensity;
-    conservation.demand_reduction_pct = calculate_demand_reduction(&conservation, precipitation);
-
-    // 2. Sewage reduction
-    conservation.sewage_reduction_pct = if conservation.greywater_recycling {
-        GREYWATER_SEWAGE_REDUCTION
-    } else {
-        0.0
-    };
-
-    // 3. Retrofit costs
-    conservation.buildings_retrofitted = building_count;
-    conservation.total_retrofit_cost = calculate_retrofit_cost(&conservation, building_count);
-
-    // 4. Annual savings estimate
-    conservation.annual_savings_gallons =
-        calculate_annual_savings(conservation.demand_reduction_pct, building_count);
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::water_conservation::calculations::*;
+    use crate::water_conservation::constants::*;
+    use crate::water_conservation::types::WaterConservationState;
 
     /// Helper: create a default conservation state with all policies off.
     fn default_state() -> WaterConservationState {
@@ -623,18 +413,5 @@ mod tests {
             + TIERED_PRICING_DEMAND_REDUCTION
             + GREYWATER_DEMAND_REDUCTION;
         assert!((sum - MAX_TOTAL_DEMAND_REDUCTION).abs() < f32::EPSILON);
-    }
-}
-
-pub struct WaterConservationPlugin;
-
-impl Plugin for WaterConservationPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<WaterConservationState>().add_systems(
-            FixedUpdate,
-            update_water_conservation
-                .after(crate::imports_exports::process_trade)
-                .in_set(crate::SimulationSet::Simulation),
-        );
     }
 }
